+++ test/color/test_adjust.py
@@ -3,7 +3,6 @@ import pytest
 import kornia
 import kornia.testing as utils  # test utils
 from kornia.constants import pi
-from test.common import device
 
 import torch
 from torch.autograd import gradcheck
