+++ test/color/test_core.py
@@ -3,7 +3,6 @@ import random
 
 import kornia
 import kornia.testing as utils  # test utils
-from test.common import device
 
 import torch
 from torch.autograd import gradcheck
