+++ test/color/test_hsv.py
@@ -3,7 +3,6 @@ import pytest
 import kornia
 from kornia.constants import pi
 import kornia.testing as utils  # test utils
-from test.common import device
 
 import math
