+++ test/color/test_rgb.py
@@ -1,6 +1,5 @@
 import kornia
 import kornia.testing as utils  # test utils
-from test.common import device
 
 import torch
 from torch.autograd import gradcheck
