+++ test/feature/test_affine_shape_estimator.py
@@ -1,7 +1,6 @@
 import pytest
 import kornia.testing as utils  # test utils
 import kornia
-from test.common import device
 
 from torch.testing import assert_allclose
 from torch.autograd import gradcheck
