+++ test/feature/test_local_features_orientation.py
@@ -1,7 +1,6 @@
 import pytest
 import kornia.testing as utils  # test utils
 import kornia
-from test.common import device
 
 from torch.testing import assert_allclose
 from torch.autograd import gradcheck
