+++ test/feature/test_siftdesc.py
@@ -1,6 +1,5 @@
 import pytest
 import kornia.testing as utils  # test utils
-from test.common import device
 
 from torch.testing import assert_allclose
 from torch.autograd import gradcheck
