+++ test/geometry/test_depth.py
@@ -2,7 +2,6 @@ import pytest
 
 import kornia
 import kornia.testing as utils  # test utils
-from test.common import device
 
 import torch
 from torch.autograd import gradcheck
