+++ test/geometry/transform/test_affine.py
@@ -1,7 +1,6 @@
 import pytest
 import kornia as kornia
 import kornia.testing as utils  # test utils
-from test.common import device
 
 import torch
 from torch.autograd import gradcheck
