+++ test/geometry/transform/test_crop.py
@@ -4,7 +4,6 @@ import pytest
 
 import kornia as kornia
 import kornia.testing as utils  # test utils
-from test.common import device
 
 import torch
 from torch.testing import assert_allclose
