+++ test/integration/test_focal.py
@@ -5,7 +5,6 @@ import torch
 import torch.nn as nn
 import torch.nn.functional as F
 import torch.optim as optim
-from test.common import device
 
 import kornia
