+++ test/integration/test_warp.py
@@ -7,7 +7,6 @@ import torch
 import torch.nn as nn
 import torch.nn.functional as F
 import torch.optim as optim
-from test.common import device
 
 
 class MyHomography(nn.Module):
