+++ test/test_losses.py
@@ -2,7 +2,6 @@ import pytest
 
 import kornia
 import kornia.testing as utils  # test utils
-from test.common import device, dtype
 
 import math
 import torch
